Ensured that --with-stdlib=libc++ is used by Homebrew and MacPorts.
Building with Clang is not sufficient in and of itself on Mac OS X,
as it defaults to using libstdc++ which at least on Mac OS X 10.7 is
still altogether missing basic C++11 headers such as <system_error>.

This also entails Mac OS X 10.7 as the minimum system requirement on
the Mac. Users on older releases of Mac OS X will need to first
install libc++ (and possibly Clang) before they will be able to build
libcpr.